Single Model Loader for all "model" modules, fixes, cleanup.
Refactor all "model" modules so they are named model_<TYPE>.
Add a modelloader class to handle fetching existing and/or models via URL.
Ensure *all* model classes aside from the main SD class *only* load/fetch models when they are actually used.
Ensure *all* model classes have a default model they can download, requiring no user intervention, period.
Move all downloaded/user model files to the /Models directory.
Update gitignore.
Fix various typos.
Fix error when user tries to upscale "nothing".
Add global "models_path" variable to paths
